Passing spec for recordArtworkView component
